app/requirements.txt
Flask==3.0.2
openai==0.28.1
python-dotenv==1.0.1
gunicorn==21.2.0
flask-cors==4.0.0
